fix: add perform Tauri migration script to v2
