+++ package.json
@@ -37,7 +37,13 @@
     "@radix-ui/react-tooltip": "^1.0.5",
     "@reduxjs/toolkit": "^1.8.6",
     "@tanstack/react-table": "8.7.9",
-    "@tauri-apps/api": "^1.6.0",
+    "@tauri-apps/api": "^2.5.0",
+    "@tauri-apps/plugin-clipboard-manager": "~2",
+    "@tauri-apps/plugin-dialog": "~2",
+    "@tauri-apps/plugin-http": "~2",
+    "@tauri-apps/plugin-notification": "~2",
+    "@tauri-apps/plugin-shell": "~2",
+    "@tauri-apps/plugin-store": "~2",
     "base64-js": "^1.5.1",
     "chart.js": "^4.2.1",
     "deck.gl": "^8.9.7",
@@ -72,7 +78,6 @@
     "redux-logger": "^3.0.6",
     "tailwindcss-radix": "^2.8.0",
     "tauri-plugin-log-api": "github:tauri-apps/tauri-plugin-log#v1",
-    "tauri-plugin-store-api": "github:tauri-apps/tauri-plugin-store#v1",
     "timeago-react": "^3.0.5",
     "tsparticles": "^2.7.1",
     "tsparticles-engine": "^2.7.1",
@@ -80,7 +85,7 @@
   },
   "devDependencies": {
     "@biomejs/biome": "~1.4.1",
-    "@tauri-apps/cli": "^1.6.3",
+    "@tauri-apps/cli": "^2.5.0",
     "@types/jest": "^29.2.0",
     "@types/lodash.clonedeep": "^4.5.7",
     "@types/lodash.debounce": "^4.0.7",
