+++ src-tauri/Cargo.toml
@@ -12,7 +12,7 @@ rust-version = "1.57"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "1.1.1", features = [] }
+tauri-build = { version = "2", features = [] }
 
 [dependencies]
 rand = "0.8.5"
@@ -22,7 +22,7 @@ defaultdict = "0.13.0"
 reqwest = { version = "0.11", features = ["json"] }
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.6.0", features = ["cli", "clipboard-write-text", "dialog-message", "http-all", "notification-all", "path-all", "shell-open", "test", "windows7-compat"] }
+tauri = { version = "2", features = [ "test"] }
 tokio = { version = "1.21.2", features = ["full"] }
 tracing = "0.1.37"
 tracing-subscriber = "0.3.16"
@@ -34,12 +34,17 @@ geojson = "0.24.0"
 log = "0.4.20"
 humantime = "2.1.0"
 tokio-util = "0.7.7"
-tauri-plugin-store = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
+tauri-plugin-store = "2"
 tokio-serial = "5.4.4"
-tauri-plugin-log = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1", features = ["colored"] }
+tauri-plugin-log = { features = ["colored"] , version = "2" }
 chrono = { version = "0.4.34", features = ["serde"] }
 meshtastic = { version = "0.1.6", features = ["ts-gen"] }
 specta = { git = "https://github.com/ajmcquilkin/specta.git", rev = "6a8731d168376e28e163dd9cd328055b11d1af82", version = "1.0.3", features = ["chrono"] }
+tauri-plugin-dialog = "2"
+tauri-plugin-shell = "2"
+tauri-plugin-notification = "2"
+tauri-plugin-clipboard-manager = "2"
+tauri-plugin-http = "2"
 
 [features]
 # by default Tauri runs in production mode
@@ -48,3 +53,7 @@ default = ["custom-protocol"]
 # this feature is used for production builds where `devPath` points to the filesystem
 # DO NOT remove this
 custom-protocol = ["tauri/custom-protocol"]
+
+[lib]
+name = "app_lib"
+crate-type = ["staticlib", "cdylib", "rlib", "lib"]
