+++ src/store/persistence.ts
@@ -1,4 +1,4 @@
-import { Store } from "tauri-plugin-store-api";
+import { Store } from "@tauri-apps/plugin-store";
 import { DEFAULT_STORE_FILE_NAME } from "@utils/persistence";
 
 export const defaultStore = new Store(DEFAULT_STORE_FILE_NAME);
