+++ src/utils/clipboard.ts
@@ -1,4 +1,4 @@
-import { writeText } from "@tauri-apps/api/clipboard";
+import { writeText } from "@tauri-apps/plugin-clipboard-manager";
 
 export const writeValueToClipboard = async (value: string) => {
   await writeText(value);
